タスク、フォルダ情報をユーザーと紐付け

+++ lib/Model/Folder.php
@@ -4,7 +4,8 @@ namespace MyApp\Model;
 class Folder extends \MyApp\Model {
 
   public function getAll() {
-    $sql = 'SELECT * FROM folder ORDER BY id ASC';
+    $user_id = $this->getUserId();
+    $sql = "SELECT * FROM folder WHERE user_id = '$user_id' ORDER BY id ASC";
     $stmt = $this->db->prepare($sql);
     $stmt->execute();
     $res = $stmt->fetchAll();
@@ -12,9 +13,10 @@ class Folder extends \MyApp\Model {
   }
 
   public function getFolder() {
+  $user_id = $this->getUserId();
   $folder_id = h($_GET['folder_id']);
 
-  $sql = "SELECT id,title FROM folder WHERE id = '$folder_id'";
+  $sql = "SELECT id,title FROM folder WHERE id = '$folder_id' AND user_id = '$user_id'";
   $stmt = $this->db->prepare($sql);
   $stmt->execute();
   $res = $stmt->fetchAll();
@@ -22,23 +24,26 @@ class Folder extends \MyApp\Model {
   }
 
   public function update($values) {
-    $sql = 'UPDATE folder SET title=? WHERE id=?';
+    $sql = 'UPDATE folder SET title=? WHERE id=? AND user_id=?';
     $data[] = $values['title'];
     $data[] = $values['folder_id'];
+    $data[] = $values['user_id'];
     $stmt = $this->db->prepare($sql);
     $stmt->execute($data);
   }
 
   public function create($values) {
-    $sql = 'INSERT INTO folder(title) VALUES (?)';
+    $sql = 'INSERT INTO folder(user_id,title) VALUES (?,?)';
     $stmt = $this->db->prepare($sql);
+    $data[] = $values['user_id'];
     $data[] = $values['title'];
     $stmt->execute($data);
   }
 
   public function delete($values) {
-    $sql = 'DELETE FROM folder WHERE id=?';
+    $sql = 'DELETE FROM folder WHERE id=? AND user_id=?';
     $data[] = $values['folder_id'];
+    $data[] = $values['user_id'];
     $stmt = $this->db->prepare($sql);
     $stmt->execute($data);
   }
